+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_advice/structure/CommandAdviceEntry.java
@@ -16,6 +16,8 @@ import lombok.ToString;
 @AllArgsConstructor
 public class CommandAdviceEntry {
 
+    boolean enable = true;
+
     Matcher matcher = new Matcher();
     @Data
     @NoArgsConstructor